+++ mods/BrewLAN/modules/BrewGantry3/mod_info.lua
@@ -1,7 +1,7 @@
 name = "BrewLAN: AIx Gantry Hax C"
-uid = "25D57D85-7D84-27HT-A503-BR3WGA3000002"
-version = 0.2
-copyright = "© 2015–2018 Sean Wheeldon"
+uid = "25D57D85-7D84-27HT-A503-BR3WGA3000003"
+version = 0.3
+copyright = "© 2015–2021 Sean Wheeldon"
 description = "Sorian AIx Gantry superhax varient 3. Mass-gated exponential growth, with discounts instead of free resources."
 author = "Balthazar and mr.deagle"
 icon = "/mods/BrewLAN/modules/BrewGantry1/mod_icon.dds"
@@ -10,6 +10,7 @@ enabled = true
 exclusive = false
 ui_only = false
 conflicts = {
-    "25D57D85-7D84-27HT-A501-BR3WGA1000002",
-    "25D57D85-7D84-27HT-A501-BR3WGA2000002",
+    "25D57D85-7D84-27HT-A501-BR3WGA1000003",
+    "25D57D85-7D84-27HT-A502-BR3WGA2000003",
 }
+after = { "25D57D85-7D84-27HT-A501-BR3WL4N0000891" }